+++ button/demo/stories.ts
@@ -120,11 +120,13 @@ const buttons: MaterialStoryInit<StoryKnobs> = {
 const links: MaterialStoryInit<StoryKnobs> = {
   name: 'Links',
   styles,
-  render({label}) {
+  render({label, disabled, softDisabled}) {
     return html`
       <div class="column">
         <div class="row">
           <md-filled-button
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -132,6 +134,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
           </md-filled-button>
 
           <md-outlined-button
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -139,6 +143,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
           </md-outlined-button>
 
           <md-elevated-button
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -146,6 +152,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
           </md-elevated-button>
 
           <md-filled-tonal-button
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -153,6 +161,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
           </md-filled-tonal-button>
 
           <md-text-button
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -162,6 +172,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
         <div class="row">
           <md-filled-button
             aria-label="${label || 'Filled'} link with trailing icon"
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -171,6 +183,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
 
           <md-outlined-button
             aria-label="${label || 'Outlined'} link with trailing icon"
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -180,6 +194,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
 
           <md-elevated-button
             aria-label="${label || 'Elevated'} link with trailing icon"
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -189,6 +205,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
 
           <md-filled-tonal-button
             aria-label="${label || 'Tonal'} link with trailing icon"
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>
@@ -198,6 +216,8 @@ const links: MaterialStoryInit<StoryKnobs> = {
 
           <md-text-button
             aria-label="${label || 'Text'} link with trailing icon"
+            ?disabled=${disabled}
+            ?soft-disabled=${softDisabled}
             href="https://google.com"
             target="_blank"
             trailing-icon>