+++ radio/internal/single-selection-controller.ts
@@ -74,27 +74,40 @@ export class SingleSelectionController implements ReactiveController {
   constructor(private readonly host: SingleSelectionElement) {}
 
   hostConnected() {
-    this.root = this.host.getRootNode() as ParentNode;
     this.host.addEventListener('keydown', this.handleKeyDown);
     this.host.addEventListener('focusin', this.handleFocusIn);
     this.host.addEventListener('focusout', this.handleFocusOut);
-    if (this.host.checked) {
-      // Uncheck other siblings when attached if already checked. This mimics
-      // native <input type="radio"> behavior.
-      this.uncheckSiblings();
-    }
 
-    // Update for the newly added host.
-    this.updateTabIndices();
+    // Update siblings after a microtask to allow other synchronous connected
+    // callbacks to settle before triggering additional Lit updates. This avoids
+    // stack overflow issues when too many elements are being rendered and
+    // connected at the same time.
+    queueMicrotask(() => {
+      // Update for the newly added host.
+      this.root = this.host.getRootNode() as ParentNode;
+      if (this.host.checked) {
+        // Uncheck other siblings when attached if already checked. This mimics
+        // native <input type="radio"> behavior.
+        this.uncheckSiblings();
+      }
+
+      this.updateTabIndices();
+    });
   }
 
   hostDisconnected() {
     this.host.removeEventListener('keydown', this.handleKeyDown);
     this.host.removeEventListener('focusin', this.handleFocusIn);
     this.host.removeEventListener('focusout', this.handleFocusOut);
-    // Update for siblings that are still connected.
-    this.updateTabIndices();
-    this.root = null;
+    // Update siblings after a microtask to allow other synchronous disconnected
+    // callbacks to settle before triggering additional Lit updates. This avoids
+    // stack overflow issues when too many elements are being rendered and
+    // connected at the same time.
+    queueMicrotask(() => {
+      // Update for siblings that are still connected.
+      this.updateTabIndices();
+      this.root = null;
+    });
   }
 
   /**