+++ select/internal/select.ts
@@ -311,6 +311,11 @@ export abstract class Select extends selectBaseClass {
     this.nativeErrorText = '';
   }
 
+  /** Shows the picker. If it's already open, this is a no-op. */
+  showPicker() {
+    this.open = true;
+  }
+
   override [onReportValidity](invalidEvent: Event | null) {
     // Prevent default pop-up behavior.
     invalidEvent?.preventDefault();
@@ -386,6 +391,7 @@ export abstract class Select extends selectBaseClass {
   }
 
   private renderField() {
+    const ariaLabel = (this as ARIAMixinStrict).ariaLabel || this.label;
     return staticHtml`
       <${this.fieldTag}
           aria-haspopup="listbox"
@@ -393,7 +399,7 @@ export abstract class Select extends selectBaseClass {
           part="field"
           id="field"
           tabindex=${this.disabled ? '-1' : '0'}
-          aria-label=${(this as ARIAMixinStrict).ariaLabel || nothing}
+          aria-label=${ariaLabel || nothing}
           aria-describedby="description"
           aria-expanded=${this.open ? 'true' : 'false'}
           aria-controls="listbox"